drm/i915/mst: Do not retrain new links
The shortpulse handler checks if channel equalization is okay and
goes onto retrain a link if there were active links existing. This
retraining path is not meant for new MST connections, but due to a bug
elsewhere, if active_link is < 0, the boolean check for active_links
proceeds to retrain a new link. This results in a sequence of failed link
training attempts, most likely due to the hardware not setup for link
training i.e., missing the DDI pre_enable sequence.

[   80.301272] [drm:intel_dp_check_mst_status] channel EQ not ok, retraining
[   80.301312] [drm:intel_ddi_prepare_link_retrain] *ERROR* Timeout waiting for DDI BUF C idle bit

The above error gives us a hint something went wrong before link
training started.

While the modeset sequence that causes this should be fixed, throw in a
warning for invalid active_mst_links for now.

Cc: Nathan Ciobanu <nathan.d.ciobanu@linux.intel.com>
Cc: Rodrigo Vivi <rodrigo.vivi@intel.com>
Signed-off-by: Dhinakaran Pandiyan <dhinakaran.pandiyan@intel.com>